requirements.txt
tqdm
wandb
numpy
matplotlib
Pillow
scikit-image
pycocotools
jsonlines
rasterio
wget==3.2
einops  